+++ packages/feedback/tsconfig.json
@@ -13,5 +13,5 @@
       "react-dom": ["./node_modules/preact/compat/"]
     }
   },
-  "include": ["src/**/*.ts","src/**/*.tsx"]
+  "include": ["src/**/*.ts", "src/**/*.tsx"]
 }
